requirements.txt
amqp==2.2.2
beautifulsoup4==4.6.0
billiard==3.5.0.3
bottlenose==1.1.8
cairocffi==0.8.0
CairoSVG==2.1.3
celery==4.1.0
cffi==1.11.5
cssselect==1.0.3
cssselect2==0.2.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.9.13
django-amazon-price-monitor==0.7
djangorestframework==3.7.7
gunicorn==19.7.1
kombu==4.1.0
lxml==4.1.1
Pillow==5.0.0
psycopg2==2.7.4
pycparser==2.18
pygal==2.4.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2018.3
six==1.11.0
tinycss==0.4
tinycss2==0.6.1
vine==1.1.4
webencodings==0.5.1
whitenoise==3.3.1
